plugins: pass down resource attributes prefixed with docker.cli to plugins
This allows `docker.cli` attributes set on the resource to be passed
down to plugins unchanged.

This also adds `docker.cli.current_context` as a resource attribute that
is included by default. The overall intention of this is to pass this
information down to compose so it doesn't have to get the information
itself and then migrate compose to use the cli package instead of its
own custom tracer provider, but the current context can be useful for
other workloads and telemetry too.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>